largest ixzc dimension
ixzc was set =100 since a time where we computed models with 300 layers only. They now have more than 1000 layers usually, so I increased the dimension for the arrays storing the number of conv. zones. In the .g file, if one radiative layer is there between 2 convective ones, it is not recorded.