Commit to add the computation and print of k2
In this commit I add the computation of the apsidal motion constant. I used a fourth order Runge Kutta method with a cubic polynomial interpolation to preserve the order of convergence. I add also added k2 in the prints (new column of the .g files).

At this point the computation of k2 still sometimes provides "unstable results" (the long term evolution is okay but there are couple outsiders, where just for one or two timesteps the value of k_AMC is too high). This will be fixed in a following commit.

I already do a commit with this version of the computation of k2 as I need this quantity to implement the equilibrium tides. But the calculation of the equilibrium tides is independent of the improvement of the computation of k2, which can therefore be done afterwards.