Updated Makefile
The git commit and compilation date are stored in the variables
GIT_COMMIT and COMPILATION_DATE.
These variables are set as preprocessor variables through the -D
compilation option. To do so, we also need to add the -cpp option
(C preprocessor).
These two variables are then available in the source code.