Commit adding equilibrium tides during MS
In this commit I add the contribution of equilibrium tides acting on small convective zones during the main squence. I used a formalism very close to that in Fragos+ 23, which is adapted from Hut81, Rasio+96, Hurley+02. The relevant equation is Eq. (6) in Fragos+23. The main difference between their approach and mine is that I obtain the convective turnover timescale self-consistently, within the MLT framework, this quantity being available in GENEC in small convective zones present in the envelope during the MS.

In this commit I also modified inputparam to add two options in the parameter card: "include_dyn_tides" and "include_eq_tides", allowing to independently check the impact of each type of tides. By default, include_dyn_tides is true, include_eq_tides is false, consistently with the default version.

In order to obtain the local values of the convective turnover timescale inside the tides module, I add them in the big array envel at the 20th entry (previously only the 19 first were given a value). tau_conv (variable tm) was previously computed in the module envelope.f90 (line 1245, now 1248) but not stored in envel.