Fix the "unstable results" of compute_k2_from_structure
In this commit I fix the bug mentionned in a previous commit where the computation of the apsidal motion constant would provide "unstable results".

This was due to the fact that when creating the arrays assembling the envelope and the interior, in some instances the built radius would not have monotonous values, which would imply an incorrect value for k2. There would be some instances where at the very bottom of the envelope, near the transition to the interior, a few layers would have a radius higher than the previous layers (in principle the envelope is stored from surface to bottom, as the interior). Removing these few layers fixed the issue.

I also updated the subroutine remove_duplicate_elements as to ensure that both the radius array and the mass array are strictly monotonous (this update also aims at improving the computed value of k2).